+++ src/codeql.test.ts
@@ -2,8 +2,8 @@ import * as fs from "fs";
 
 import { ExecOptions } from "@actions/exec";
 import * as toolrunner from "@actions/exec/lib/toolrunner";
+import * as io from "@actions/io";
 import * as toolcache from "@actions/tool-cache";
-import * as safeWhich from "@chrisgavin/safe-which";
 import test, { ExecutionContext } from "ava";
 import del from "del";
 import * as yaml from "js-yaml";
@@ -791,14 +791,6 @@ const NEW_ANALYSIS_SUMMARY_TEST_CASES = [
     flagPassed: true,
     negativeFlagPassed: false,
   },
-  {
-    codeqlVersion: makeVersionInfo("2.14.6"),
-    githubVersion: {
-      type: util.GitHubVariant.DOTCOM,
-    },
-    flagPassed: false,
-    negativeFlagPassed: false,
-  },
 ];
 
 for (const {
@@ -819,8 +811,8 @@ for (const {
     const runnerConstructorStub = stubToolRunnerConstructor();
     const codeqlObject = await codeql.getCodeQLForTesting();
     sinon.stub(codeqlObject, "getVersion").resolves(codeqlVersion);
-    // safeWhich throws because of the test CodeQL object.
-    sinon.stub(safeWhich, "safeWhich").resolves("");
+    // io throws because of the test CodeQL object.
+    sinon.stub(io, "which").resolves("");
     await codeqlObject.databaseInterpretResults(
       "",
       [],
@@ -860,8 +852,8 @@ test("runTool summarizes several fatal errors", async (t) => {
   stubToolRunnerConstructor(32, cliStderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
   sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
+  // io throws because of the test CodeQL object.
+  sinon.stub(io, "which").resolves("");
 
   await t.throwsAsync(
     async () =>
@@ -902,8 +894,8 @@ test("runTool summarizes autobuilder errors", async (t) => {
   const codeqlObject = await codeql.getCodeQLForTesting();
   sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   sinon.stub(codeqlObject, "resolveExtractor").resolves("/path/to/extractor");
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
+  // io throws because of the test CodeQL object.
+  sinon.stub(io, "which").resolves("");
 
   await t.throwsAsync(
     async () => await codeqlObject.runAutobuild(stubConfig, Language.java),
@@ -929,8 +921,8 @@ test("runTool truncates long autobuilder errors", async (t) => {
   const codeqlObject = await codeql.getCodeQLForTesting();
   sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   sinon.stub(codeqlObject, "resolveExtractor").resolves("/path/to/extractor");
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
+  // io throws because of the test CodeQL object.
+  sinon.stub(io, "which").resolves("");
 
   await t.throwsAsync(
     async () => await codeqlObject.runAutobuild(stubConfig, Language.java),
@@ -957,15 +949,15 @@ test("runTool recognizes fatal internal errors", async (t) => {
   const codeqlObject = await codeql.getCodeQLForTesting();
   sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   sinon.stub(codeqlObject, "resolveExtractor").resolves("/path/to/extractor");
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
+  // io throws because of the test CodeQL object.
+  sinon.stub(io, "which").resolves("");
 
   await t.throwsAsync(
     async () =>
       await codeqlObject.databaseRunQueries(stubConfig.dbLocation, []),
     {
       instanceOf: CliError,
-      message: `Encountered a fatal error while running "codeql-for-testing database run-queries  --expect-discarded-cache --min-disk-free=1024 -v --intra-layer-parallelism". Exit code was 1 and error was: Oops! A fatal internal error occurred. Details:
+      message: `Encountered a fatal error while running "codeql-for-testing database run-queries  --expect-discarded-cache --intra-layer-parallelism --min-disk-free=1024 -v". Exit code was 1 and error was: Oops! A fatal internal error occurred. Details:
     com.semmle.util.exception.CatastrophicError: An error occurred while evaluating ControlFlowGraph::ControlFlow::Root.isRootOf/1#dispred#f610e6ed/2@86282cc8
     Severe disk cache trouble (corruption or out of space) at /home/runner/work/_temp/codeql_databases/go/db-go/default/cache/pages/28/33.pack: Failed to write item to disk. See the logs for more details.`,
     },
@@ -977,8 +969,8 @@ test("runTool outputs last line of stderr if fatal error could not be found", as
   stubToolRunnerConstructor(32, cliStderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
   sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
+  // io throws because of the test CodeQL object.
+  sinon.stub(io, "which").resolves("");
 
   await t.throwsAsync(
     async () =>
@@ -1002,8 +994,8 @@ test("Avoids duplicating --overwrite flag if specified in CODEQL_ACTION_EXTRA_OP
   const runnerConstructorStub = stubToolRunnerConstructor();
   const codeqlObject = await codeql.getCodeQLForTesting();
   sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
+  // io throws because of the test CodeQL object.
+  sinon.stub(io, "which").resolves("");
 
   process.env["CODEQL_ACTION_EXTRA_OPTIONS"] =
     '{ "database": { "init": ["--overwrite"] } }';